requirements.txt
amqp==5.3.1
asgiref==3.7.2
async-timeout==5.0.1
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.5
django-cors-headers==4.2.0
django-jazzmin==3.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
kombu==5.4.2
prompt_toolkit==3.0.50
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2023.3.post1
redis==5.2.1
six==1.17.0
sqlparse==0.4.4
typing_extensions==4.12.2
tzdata==2023.3
vine==5.1.0
wcwidth==0.2.13
